[DP-163] fix: BigDecimal 타입 변경에 따른 수정 #129

+++ src/test/java/com/dapanda/TestConstants.java
@@ -16,8 +16,8 @@ public final class TestConstants {
 		public static final Long OTHER_MEMBER_ID = 2L;
 		public static final int CASH_3000 = 3000;
 		public static final int CASH_5000 = 5000;
-		public static final float BUYING_DATA = 1F;
-		public static final float SELLING_DATA = 1F;
+		public static final BigDecimal BUYING_DATA = BigDecimal.valueOf(1.0);
+		public static final BigDecimal SELLING_DATA = BigDecimal.valueOf(1.0);
 	}
 
 	public static final class Pagination {
